+++ src/models/IStatementGiver.ts
@@ -1,0 +1,5 @@
+import { BuilderData } from '../builder'
+
+export interface IStatementGiver {
+  getStmt(data: BuilderData): string
+}